auto configure egress
