+++ logic/acls.go
@@ -223,76 +223,6 @@ func IsAclExists(aclID string) bool {
 	_, err := GetAcl(aclID)
 	return err == nil
 }
-func GetEgressRanges(netID models.NetworkID) (map[string][]string, map[string]struct{}, error) {
-
-	resultMap := make(map[string]struct{})
-	nodeEgressMap := make(map[string][]string)
-	networkNodes, err := GetNetworkNodes(netID.String())
-	if err != nil {
-		return nil, nil, err
-	}
-	for _, currentNode := range networkNodes {
-		if currentNode.Network != netID.String() {
-			continue
-		}
-		if currentNode.IsEgressGateway { // add the egress gateway range(s) to the result
-			if len(currentNode.EgressGatewayRanges) > 0 {
-				nodeEgressMap[currentNode.ID.String()] = currentNode.EgressGatewayRanges
-				for _, egressRangeI := range currentNode.EgressGatewayRanges {
-					resultMap[egressRangeI] = struct{}{}
-				}
-			}
-		}
-	}
-	extclients, _ := GetNetworkExtClients(netID.String())
-	for _, extclient := range extclients {
-		if len(extclient.ExtraAllowedIPs) > 0 {
-			nodeEgressMap[extclient.ClientID] = extclient.ExtraAllowedIPs
-			for _, extraAllowedIP := range extclient.ExtraAllowedIPs {
-				resultMap[extraAllowedIP] = struct{}{}
-			}
-		}
-	}
-	return nodeEgressMap, resultMap, nil
-}
-func sortRouteMetricByAscending(items []models.EgressRangeMetric) []models.EgressRangeMetric {
-	sort.Slice(items, func(i, j int) bool {
-		return items[i].RouteMetric < items[j].RouteMetric
-	})
-	return items
-}
-
-func GetEgressRangesWithMetric(netID models.NetworkID) map[string][]models.EgressRangeMetric {
-
-	egressMap := make(map[string][]models.EgressRangeMetric)
-	networkNodes, err := GetNetworkNodes(netID.String())
-	if err != nil {
-		return nil
-	}
-	for _, currentNode := range networkNodes {
-		if currentNode.Network != netID.String() {
-			continue
-		}
-		if currentNode.IsEgressGateway { // add the egress gateway range(s) to the result
-			if len(currentNode.EgressGatewayRequest.RangesWithMetric) > 0 {
-				for _, egressRangeI := range currentNode.EgressGatewayRequest.RangesWithMetric {
-					if value, ok := egressMap[egressRangeI.Network]; ok {
-						value = append(value, egressRangeI)
-						egressMap[egressRangeI.Network] = value
-					} else {
-						egressMap[egressRangeI.Network] = []models.EgressRangeMetric{
-							egressRangeI,
-						}
-					}
-				}
-			}
-		}
-	}
-	for key, value := range egressMap {
-		egressMap[key] = sortRouteMetricByAscending(value)
-	}
-	return egressMap
-}
 
 func checkIfAclTagisValid(t models.AclPolicyTag, netID models.NetworkID, policyType models.AclPolicyType, isSrc bool) bool {
 	switch t.ID {
